pyproject.toml
[project]
name = "meetscribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.54",
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "langchain>=1.2.10",
    "langchain-aws>=1.2.5",
    "langchain-community>=0.4.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.54.0",
]
